Working with knex and no typeorm
