+++ src/controllers/request-controller.ts
@@ -7,10 +7,9 @@ import cors from 'cors'
 import { ClassMiddleware, Controller, Get, Post } from '@overnightjs/core'
 
 import { toCID } from '@ceramicnetwork/common'
-import { RequestStatus } from '../models/request-status.js'
 import { AnchorRepository } from '../repositories/anchor-repository.js'
 import { RequestRepository } from '../repositories/request-repository.js'
-import { Request } from '../models/request.js'
+import { Request, RequestStatus } from '../models/request.js'
 import { inject, singleton } from 'tsyringe'
 import { logger } from '../logger/index.js'
 import { RequestPresentation } from '../models/request-presentation.js'
