--- src/models/request-status.ts
@@ -1,13 +1,0 @@
-export enum RequestStatus {
-  PENDING = 0,
-  PROCESSING = 1,
-  COMPLETED = 2,
-  FAILED = 3,
-  READY = 4,
-}
-
-export class InvalidRequestStatusError extends Error {
-  constructor(status: never) {
-    super(`Invalid request status: ${status}`)
-  }
-}
