+++ src/models/request.ts
@@ -1,39 +1,24 @@
-import { RequestStatus } from './request-status.js'
-import TypeORM from 'typeorm'
-const { Entity, PrimaryGeneratedColumn, Column, CreateDateColumn, UpdateDateColumn, Unique } =
-  TypeORM
+export const TABLE_NAME = 'request'
 
-@Entity()
-@Unique(['cid'])
+export enum RequestStatus {
+  PENDING = 0,
+  PROCESSING = 1,
+  COMPLETED = 2,
+  FAILED = 3,
+  READY = 4,
+}
+
 export class Request {
-  @PrimaryGeneratedColumn('uuid')
   id: number
-
-  @Column({ nullable: false })
   status: RequestStatus
-
-  @Column({ nullable: false })
   cid: string
-
-  @Column({ nullable: false, name: 'doc_id' })
   streamId: string
-
-  @Column({ nullable: true })
   message: string
-
-  @Column({ nullable: false, default: false })
   pinned: boolean
-
-  @CreateDateColumn({ name: 'created_at' })
   createdAt: Date
-
-  @UpdateDateColumn({ name: 'updated_at' })
   updatedAt: Date
 }
 
-/**
- * Request update  fields
- */
 export interface RequestUpdateFields {
   message?: string
   status?: RequestStatus
@@ -42,4 +27,10 @@ export interface RequestUpdateFields {
 
 export const REQUEST_MESSAGES = {
   conflictResolutionRejection: 'Request has failed. Updated was rejected by conflict resolution.',
+}
+
+export class InvalidRequestStatusError extends Error {
+  constructor(status: never) {
+    super(`Invalid request status: ${status}`)
+  }
 }
